requirements.txt
uvicorn
fastapi
numpy
pandas
ipaddress
requests
python-whois
urllib3
xgboost
scikit-learn